Hacky demo2 with OO paradigm + no Js.Unsafe. exposed (hopefully...)
It's hard to make sure that you type checked things properly when
using jsoo because all you need is to have a `: any` and the type
checker pretty much bails on you (and it errors are runtime).
I tried to wrap all of the unsafe calls with type-safe APIs, but I'm
not sure I did it properly.